third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/no-form-association-expected.txt

Form control elements inside templates should not be associated with forms outside the template

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Form in document, input inside template:
PASS form.length is 0
PASS input.form is null

Form in template, input in sub-template:
PASS form.length is 0
PASS input.form is null
PASS successfullyParsed is true

TEST COMPLETE
